fixed all posts and authors endpoints in the new format
