+++ distributed_social_network/api/views/__init__.py
@@ -1,2 +1,3 @@
-from .views import *
 from .followers import *
+from .posts import *
+from .authors import *
