+++ distributed_social_network/api/views/posts.py
@@ -2,12 +2,40 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.http import JsonResponse, HttpResponse
 from ..serializers import PostSerializer
 import base64
+from ..models import Post
+from rest_framework.decorators import api_view
+
+# Routes the request for a single post
+@api_view(['GET', 'POST', 'DELETE', 'PUT'])
+def route_single_post(request, author_id, post_id):
+    if request.method == 'GET':
+        return get_post(request, post_id)
+    elif request.method == 'POST':
+        return update_post(request, post_id)
+    elif request.method == 'DELETE':
+        return delete_post(request, post_id)
+    elif request.method == 'PUT':
+        return create_post(request, post_id)
+
+# Routes the request for multiple posts
+@api_view(['POST', 'GET'])
+def route_multiple_posts(request, author_id):
+    if request.method == 'POST':
+        return create_post(request, post_id)
+    elif request.method == 'GET':
+        return get_multiple_posts(request)
+
+# Routes the request for a single image post
+@api_view(['GET'])
+def route_single_image_post(request, author_id, post_id):
+    if request.method == 'GET':
+        return get_post_image(request, post_id)
 
 # Adds a new post to the database.
 # Expects JSON request body with post attributes.
-def create_post(request, post_id):
+def create_post(request, id):
     # Use the given id
-    request.data["id"] = post_id
+    request.data["id"] = id
     
     # Serialize a new Post object
     serializer = PostSerializer(data = request.data)
@@ -44,6 +72,9 @@ def delete_post(request, id):
 # Updates the post with id 'id' in the database.
 def update_post(request, id):
     response = HttpResponse()
+
+    # Use the given id
+    request.data["id"] = id
 
     # Find the post with the given id
     post = find_post(id)
